This commit introduces the electron autoupdater to the zen wallet to enable updates to the app be more seamless and better integrated with electron.
It introduces a checkbox item in the settings pane that enables or
disables automatic checks for updates. If this toggle is on, then a
check is made to the release provider (github) every 5 minutes. If the
toggle is off, then a button is rendered that allows the user to
manually trigger a check for updates.

If there is an update, a modal is displayed to the user to ask if they
would like to download the udpate, skip this version or ask later. If
the user elects to download the update, then the updater downloads the
update in the background and notifies the user when the download is
ready to be installed. In this case the user has the option to quit the
app and install the update or close the modal. Note that once an update has been downloaded,
electron will inevitably install it when next the app is restarted.
If the user elects to skip the proposed update, then they will not be
prompted to download an update until a different version of the app has
been published. Finally if the user chooses the `ask later` option, then
they will be reminded about an available update (approximately 24hrs
later).

For debugging, this commit also introduces electron-log, which easy app
logging to a file. This was incredibly useful for building out the
feature.

The following are still outstanding:
1. The release provider still does not provide a `latest-mac.yml` and a
`latest.yml` file. Instead, there is a `latest-mac.json`. The
electron-updater requires the yml files.

2. This needs to be tested on windows.