attempt to get dapp-communication to work
I tried producing a packaged app locally on ubuntu `18.04.01` and test this, and it didn't work. It might be becuase our app isn't installed, but only packaged and run directly through the unzipped executable.

To test on linux
1. produce an installer
1. install the app
1. run the app at least once
1. close app
1. run `$ npm run dapp-communication-example`
1. click on the link

It should open the app and prefill the input field.

Code in this commit was following [this SO thread](https://stackoverflow.com/questions/53530726/how-to-launch-my-electron-app-from-a-website).